+++ src/pages/ProductList/ProductList.jsx
@@ -1,16 +1,16 @@
 import { ProductDataList, SideBar } from "./index";
-import {Navbar} from "../../Components/Navbar/Navbar";
-import { Button } from "../../Components/Buttons/Button";
+import { Button, Navbar } from "../../Components/index";
+
 
 export const ProductList = () => {
     return (
         <div className="productList-temp grid">
-                <Navbar />
-                <SideBar />
-                <ProductDataList />
-        <footer class="filter-hidden display-none align-items">
-            <Button className="xsm-heading font-weight-5 bg-transparent btn-border-none" text="Filter"/>
-        </footer>
-    </div>
+            <Navbar />
+            <SideBar />
+            <ProductDataList />
+            <footer className="filter-hidden display-none align-items">
+                <Button className="xsm-heading font-weight-5 bg-transparent btn-border-none" text="Filter"/>
+            </footer>
+        </div>
     )
 }
